+++ libCacheSim/include/libCacheSim/cacheObj.h
@@ -137,7 +137,7 @@ struct cache_obj;
 typedef struct cache_obj {
   struct cache_obj *hash_next;
   obj_id_t obj_id;
-  uint32_t obj_size;
+  uint64_t obj_size;
   struct {
     struct cache_obj *prev;
     struct cache_obj *next;
